DeepLearningMinor/requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.1.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
imageio==2.5.0
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mpmath==1.1.0
networkx==2.3
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
pickleshare==0.7.5
Pillow==6.1.0
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sympy==1.4
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.15.4
wrapt==1.11.2